Extract mergeConfigs
